CXP-954: [rule] Each backend controller should be invokable, have the Action suffix and be final
