+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Apps/Controller/AuthorizeAction.php
@@ -27,7 +27,7 @@ use Symfony\Component\Routing\RouterInterface;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class AuthorizeAction
+final class AuthorizeAction
 {
     private RequestAppAuthorizationHandler $requestAppAuthorizationHandler;
     private RouterInterface $router;
