+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Apps/Controller/ConfirmAuthenticationAction.php
@@ -26,7 +26,7 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ConfirmAuthenticationAction
+final class ConfirmAuthenticationAction
 {
     private FeatureFlag $featureFlag;
     private GetAppConfirmationQueryInterface $getAppConfirmationQuery;
