+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Apps/Controller/ConfirmAuthorizationAction.php
@@ -29,7 +29,7 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ConfirmAuthorizationAction
+final class ConfirmAuthorizationAction
 {
     private CreateAppWithAuthorizationHandler $createAppWithAuthorizationHandler;
     private FeatureFlag $featureFlag;
