+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Apps/Controller/GetWizardDataAction.php
@@ -18,7 +18,7 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class GetWizardDataAction
+final class GetWizardDataAction
 {
     private GetAppQueryInterface $getAppQuery;
     private AppAuthorizationSessionInterface $appAuthorizationSession;
