+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Apps/Controller/InternalApi/GetConnectedAppMonitoringSettingsAction.php
@@ -19,7 +19,7 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class GetConnectedAppMonitoringSettingsAction
+final class GetConnectedAppMonitoringSettingsAction
 {
     public function __construct(
         private FeatureFlag $featureFlag,
