+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Apps/Controller/InternalApi/UpdateConnectedAppMonitoringSettingsAction.php
@@ -24,7 +24,7 @@ use Symfony\Component\Validator\ConstraintViolationListInterface;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class UpdateConnectedAppMonitoringSettingsAction
+final class UpdateConnectedAppMonitoringSettingsAction
 {
     public function __construct(
         private FeatureFlag $featureFlag,
