+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Apps/Controller/RequestAccessTokenAction.php
@@ -18,7 +18,7 @@ use Symfony\Component\Validator\Validator\ValidatorInterface;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class RequestAccessTokenAction
+final class RequestAccessTokenAction
 {
     private FeatureFlag $featureFlag;
     private ValidatorInterface $validator;
