+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/InternalApi/Controller/Connection/CreateConnectionAction.php
@@ -1,0 +1,69 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection;
+
+use Akeneo\Connectivity\Connection\Application\Settings\Command\CreateConnectionCommand;
+use Akeneo\Connectivity\Connection\Application\Settings\Command\CreateConnectionHandler;
+use Akeneo\Connectivity\Connection\Domain\Settings\Exception\ConstraintViolationListException;
+use Oro\Bundle\SecurityBundle\SecurityFacade;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Symfony\Component\Validator\ConstraintViolationListInterface;
+
+/**
+ * @author Romain Monceau <romain@akeneo.com>
+ * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
+ * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ */
+final class CreateConnectionAction
+{
+    public function __construct(
+        private CreateConnectionHandler $createConnectionHandler,
+        private SecurityFacade $securityFacade,
+    ) {
+    }
+
+    public function __invoke(Request $request): JsonResponse
+    {
+        if (true !== $this->securityFacade->isGranted('akeneo_connectivity_connection_manage_settings')) {
+            throw new AccessDeniedException();
+        }
+
+        $data = json_decode($request->getContent(), true);
+        // TODO: Valid JSON format
+
+        $command = new CreateConnectionCommand($data['code'], $data['label'], $data['flow_type'], false);
+
+        try {
+            $connection = $this->createConnectionHandler->handle($command);
+        } catch (ConstraintViolationListException $e) {
+            $errorList = $this->buildViolationResponse($e->getConstraintViolationList());
+
+            return new JsonResponse(
+                ['errors' => $errorList, 'message' => $e->getMessage()],
+                Response::HTTP_UNPROCESSABLE_ENTITY
+            );
+        } catch (\Exception $e) {
+            return new JsonResponse(['message' => $e->getMessage()], Response::HTTP_BAD_REQUEST);
+        }
+
+        return new JsonResponse($connection->normalize(), Response::HTTP_CREATED);
+    }
+
+    private function buildViolationResponse(ConstraintViolationListInterface $constraintViolationList): array
+    {
+        $errors = [];
+        foreach ($constraintViolationList as $constraintViolation) {
+            $errors[] = [
+                'name' => $constraintViolation->getPropertyPath(),
+                'reason' => $constraintViolation->getMessage(),
+            ];
+        }
+
+        return $errors;
+    }
+}
