+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/InternalApi/Controller/Connection/UpdateConnectionAction.php
@@ -1,0 +1,77 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection;
+
+use Akeneo\Connectivity\Connection\Application\Settings\Command\UpdateConnectionCommand;
+use Akeneo\Connectivity\Connection\Application\Settings\Command\UpdateConnectionHandler;
+use Akeneo\Connectivity\Connection\Domain\Settings\Exception\ConstraintViolationListException;
+use Oro\Bundle\SecurityBundle\SecurityFacade;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Symfony\Component\Validator\ConstraintViolationListInterface;
+
+/**
+ * @author Romain Monceau <romain@akeneo.com>
+ * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
+ * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ */
+final class UpdateConnectionAction
+{
+    public function __construct(
+        private UpdateConnectionHandler $updateConnectionHandler,
+        private SecurityFacade $securityFacade,
+    ) {
+    }
+
+    public function __invoke(Request $request): JsonResponse
+    {
+        if (true !== $this->securityFacade->isGranted('akeneo_connectivity_connection_manage_settings')) {
+            throw new AccessDeniedException();
+        }
+
+        $data = json_decode($request->getContent(), true);
+        // TODO: Valid JSON format
+
+        $command = new UpdateConnectionCommand(
+            $request->get('code', ''),
+            $data['label'],
+            $data['flow_type'],
+            $data['image'],
+            $data['user_role_id'],
+            $data['user_group_id'],
+            $data['auditable'] ?? false // TODO: Remove default value when auditable field will be implemented in the UI
+        );
+
+        try {
+            $this->updateConnectionHandler->handle($command);
+        } catch (ConstraintViolationListException $e) {
+            $errorList = $this->buildViolationResponse($e->getConstraintViolationList());
+
+            return new JsonResponse(
+                ['errors' => $errorList, 'message' => $e->getMessage()],
+                Response::HTTP_UNPROCESSABLE_ENTITY
+            );
+        } catch (\Exception $e) {
+            return new JsonResponse(['message' => $e->getMessage()], Response::HTTP_BAD_REQUEST);
+        }
+
+        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
+    }
+
+    private function buildViolationResponse(ConstraintViolationListInterface $constraintViolationList): array
+    {
+        $errors = [];
+        foreach ($constraintViolationList as $constraintViolation) {
+            $errors[] = [
+                'name' => $constraintViolation->getPropertyPath(),
+                'reason' => $constraintViolation->getMessage(),
+            ];
+        }
+
+        return $errors;
+    }
+}
