+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/InternalApi/Controller/EventsApiDebug/DownloadEventSubscriptionLogsAction.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller;
+namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\EventsApiDebug;
 
 use Akeneo\Connectivity\Connection\Domain\Webhook\Persistence\Query\GetAllEventSubscriptionDebugLogsQueryInterface;
-use Akeneo\Connectivity\Connection\Domain\Webhook\Persistence\Query\SearchEventSubscriptionDebugLogsQueryInterface;
 use Oro\Bundle\SecurityBundle\SecurityFacade;
 use Symfony\Component\HttpFoundation\HeaderUtils;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\StreamedResponse;
@@ -18,23 +16,15 @@ use Symfony\Component\Security\Core\Exception\AccessDeniedException;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class EventsApiDebugController
+final class DownloadEventSubscriptionLogsAction
 {
-    private GetAllEventSubscriptionDebugLogsQueryInterface $getAllEventSubscriptionDebugLogsQuery;
-    private SearchEventSubscriptionDebugLogsQueryInterface $searchEventSubscriptionDebugLogsQuery;
-    private SecurityFacade $securityFacade;
-
     public function __construct(
-        GetAllEventSubscriptionDebugLogsQueryInterface $getAllEventSubscriptionDebugLogsQuery,
-        SearchEventSubscriptionDebugLogsQueryInterface $searchEventSubscriptionDebugLogsQuery,
-        SecurityFacade $securityFacade
+        private GetAllEventSubscriptionDebugLogsQueryInterface $getAllEventSubscriptionDebugLogsQuery,
+        private SecurityFacade $securityFacade,
     ) {
-        $this->getAllEventSubscriptionDebugLogsQuery = $getAllEventSubscriptionDebugLogsQuery;
-        $this->searchEventSubscriptionDebugLogsQuery = $searchEventSubscriptionDebugLogsQuery;
-        $this->securityFacade = $securityFacade;
     }
 
-    public function downloadEventSubscriptionLogs(Request $request): Response
+    public function __invoke(Request $request): Response
     {
         if (true !== $this->securityFacade->isGranted('akeneo_connectivity_connection_manage_settings')) {
             throw new AccessDeniedException();
@@ -83,20 +73,5 @@ class EventsApiDebugController
         );
 
         return $response;
-    }
-
-    public function searchEventSubscriptionLogs(Request $request): Response
-    {
-        if (true !== $this->securityFacade->isGranted('akeneo_connectivity_connection_manage_settings')) {
-            throw new AccessDeniedException();
-        }
-
-        $connectionCode = $request->query->get('connection_code');
-        $searchAfter = $request->query->get('search_after');
-        $filters = json_decode($request->query->get('filters', ''), true) ?: [];
-
-        $logs = $this->searchEventSubscriptionDebugLogsQuery->execute($connectionCode, $searchAfter, $filters);
-
-        return new JsonResponse($logs);
     }
 }
