+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/InternalApi/Controller/EventsApiDebug/SearchEventSubscriptionLogsAction.php
@@ -1,0 +1,40 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\EventsApiDebug;
+
+use Akeneo\Connectivity\Connection\Domain\Webhook\Persistence\Query\SearchEventSubscriptionDebugLogsQueryInterface;
+use Oro\Bundle\SecurityBundle\SecurityFacade;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+
+/**
+ * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+final class SearchEventSubscriptionLogsAction
+{
+    public function __construct(
+        private SearchEventSubscriptionDebugLogsQueryInterface $searchEventSubscriptionDebugLogsQuery,
+        private SecurityFacade $securityFacade,
+    ) {
+    }
+
+    public function __invoke(Request $request): Response
+    {
+        if (true !== $this->securityFacade->isGranted('akeneo_connectivity_connection_manage_settings')) {
+            throw new AccessDeniedException();
+        }
+
+        $connectionCode = $request->query->get('connection_code');
+        $searchAfter = $request->query->get('search_after');
+        $filters = json_decode($request->query->get('filters', ''), true) ?: [];
+
+        $logs = $this->searchEventSubscriptionDebugLogsQuery->execute($connectionCode, $searchAfter, $filters);
+
+        return new JsonResponse($logs);
+    }
+}
