+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/InternalApi/Controller/Marketplace/GetAllAppsAction.php
@@ -4,8 +4,11 @@ declare(strict_types=1);
 
 namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Marketplace;
 
+use Akeneo\Connectivity\Connection\Application\Marketplace\AppUrlGenerator;
 use Akeneo\Connectivity\Connection\Application\Marketplace\MarketplaceAnalyticsGenerator;
-use Akeneo\Connectivity\Connection\Domain\Marketplace\GetAllExtensionsQueryInterface;
+use Akeneo\Connectivity\Connection\Domain\Marketplace\DTO\GetAllAppsResult;
+use Akeneo\Connectivity\Connection\Domain\Marketplace\GetAllAppsQueryInterface;
+use Akeneo\Platform\Bundle\FeatureFlagBundle\FeatureFlag;
 use Akeneo\UserManagement\Bundle\Context\UserContext;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -15,25 +18,18 @@ use Symfony\Component\HttpFoundation\Response;
 
 /**
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class GetAllExtensions
+final class GetAllAppsAction
 {
-    private GetAllExtensionsQueryInterface $getAllExtensionsQuery;
-    private MarketplaceAnalyticsGenerator $marketplaceAnalyticsGenerator;
-    private UserContext $userContext;
-    private LoggerInterface $logger;
-
     public function __construct(
-        GetAllExtensionsQueryInterface $getAllExtensionsQuery,
-        MarketplaceAnalyticsGenerator $marketplaceAnalyticsGenerator,
-        UserContext $userContext,
-        LoggerInterface $logger
+        private AppUrlGenerator $appUrlGenerator,
+        private GetAllAppsQueryInterface $getAllAppsQuery,
+        private MarketplaceAnalyticsGenerator $marketplaceAnalyticsGenerator,
+        private UserContext $userContext,
+        private LoggerInterface $logger,
+        private FeatureFlag $activateFeatureFlag,
     ) {
-        $this->getAllExtensionsQuery = $getAllExtensionsQuery;
-        $this->marketplaceAnalyticsGenerator = $marketplaceAnalyticsGenerator;
-        $this->userContext = $userContext;
-        $this->logger = $logger;
     }
 
     public function __invoke(Request $request): Response
@@ -42,10 +38,14 @@ class GetAllExtensions
             return new RedirectResponse('/');
         }
 
+        if (!$this->activateFeatureFlag->isEnabled()) {
+            return new JsonResponse(GetAllAppsResult::create(0, [])->normalize());
+        }
+
         try {
-            $result = $this->getAllExtensionsQuery->execute();
+            $result = $this->getAllAppsQuery->execute();
         } catch (\Exception $e) {
-            $this->logger->error(sprintf('unable to retrieve the list of extensions, got error "%s"', $e->getMessage()));
+            $this->logger->error(sprintf('unable to retrieve the list of apps, got error "%s"', $e->getMessage()));
 
             return new Response(null, Response::HTTP_NO_CONTENT);
         }
@@ -53,6 +53,8 @@ class GetAllExtensions
         $username = $this->userContext->getUser()->getUsername();
         $analyticsQueryParameters = $this->marketplaceAnalyticsGenerator->getExtensionQueryParameters($username);
         $result = $result->withAnalytics($analyticsQueryParameters);
+        $appQueryParameters = $this->appUrlGenerator->getAppQueryParameters();
+        $result = $result->withPimUrlSource($appQueryParameters);
 
         return new JsonResponse($result->normalize());
     }
