+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/InternalApi/Controller/Marketplace/GetWebMarketplaceUrlAction.php
@@ -16,17 +16,12 @@ use Symfony\Component\HttpFoundation\Response;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class GetWebMarketplaceUrl
+final class GetWebMarketplaceUrlAction
 {
-    private MarketplaceUrlGeneratorInterface $marketplaceUrlGenerator;
-    private UserContext $userContext;
-
     public function __construct(
-        MarketplaceUrlGeneratorInterface $marketplaceUrlGenerator,
-        UserContext $userContext
+        private MarketplaceUrlGeneratorInterface $marketplaceUrlGenerator,
+        private UserContext $userContext,
     ) {
-        $this->marketplaceUrlGenerator = $marketplaceUrlGenerator;
-        $this->userContext = $userContext;
     }
 
     public function __invoke(Request $request): Response
