+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/routing/audit.yml
@@ -1,18 +1,18 @@
 weekly:
     path: '/weekly'
-    defaults: { _controller: akeneo_connectivity.connection.internal_api.controller.audit:getWeeklyAudit }
+    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Audit\GetWeeklyAuditAction
     methods: [GET]
     requirements:
         event_type: '[a-zA-Z_]+'
 
 weekly_error:
     path: '/weekly-error'
-    defaults: { _controller: akeneo_connectivity.connection.internal_api.controller.audit:getWeeklyErrorAudit }
+    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Audit\GetWeeklyErrorAuditAction
     methods: [GET]
 
 error_count_per_connection:
     path: '/error-count-per-connection'
-    defaults: { _controller: akeneo_connectivity.connection.internal_api.controller.audit:getErrorCountPerConnection }
+    controller: Akeneo\Connectivity\Connection\Infrastructure\Apps\Controller\InternalApi\UpdateConnectedAppMonitoringSettingsAction
     methods: [GET]
     requirements:
         error_type: '[a-zA-Z_]+'
