+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/routing/connection.yml
@@ -1,49 +1,49 @@
 list:
     path: ''
-    defaults: { _controller: akeneo_connectivity.connection.internal_api.controller.connection:list }
+    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\ListConnectionsAction
     methods: [GET]
 
 create:
     path: ''
-    defaults: { _controller: akeneo_connectivity.connection.internal_api.controller.connection:create }
+    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\CreateConnectionAction
     methods: [POST]
 
 get:
     path: '/{code}'
-    defaults: { _controller: akeneo_connectivity.connection.internal_api.controller.connection:get }
+    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\GetConnectionAction
     methods: [GET]
     requirements:
         code: '[a-zA-Z0-9_]+'
 
 update:
     path: '/{code}'
-    defaults: { _controller: akeneo_connectivity.connection.internal_api.controller.connection:update }
+    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\UpdateConnectionAction
     methods: [POST]
     requirements:
         code: '[a-zA-Z0-9_]+'
 
 delete:
     path: '/{code}'
-    defaults: { _controller: akeneo_connectivity.connection.internal_api.controller.connection:delete }
+    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\DeleteConnectionAction
     methods: [DELETE]
     requirements:
         code: '[a-zA-Z0-9_]+'
 
 regenerate_secret:
     path: '/{code}/regenerate-secret'
-    defaults: { _controller: akeneo_connectivity.connection.internal_api.controller.connection:regenerateSecret }
+    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\RegenerateSecretAction
     methods: [POST]
     requirements:
         code: '[a-zA-Z0-9_]+'
 
 regenerate_password:
     path: '/{code}/regenerate-password'
-    defaults: { _controller: akeneo_connectivity.connection.internal_api.controller.connection:regeneratePassword }
+    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\RegeneratePasswordAction
     methods: [POST]
     requirements:
         code: '[a-zA-Z0-9_]+'
 
 wrong_credentials_combination_list:
     path: '/wrong-credentials-combination'
-    defaults: { _controller: akeneo_connectivity.connection.internal_api.wrong_credentials_combination:list }
+    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\WrongCredentialsCombination\ListWrongCredentialsCombinationsAction
     methods: [GET]
