+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/wrong_credentials_connection.yml
@@ -22,8 +22,7 @@ services:
         tags:
             - { name: kernel.event_subscriber }
 
-    akeneo_connectivity.connection.internal_api.wrong_credentials_combination:
+    Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\WrongCredentialsCombination\ListWrongCredentialsCombinationsAction:
         public: true
-        class: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\WrongCredentialsCombinationController
         arguments:
             - '@akeneo_connectivity.connection.persistence.repository.wrong_credentials_combination'
