+++ src/Akeneo/Connectivity/Connection/back/tests/Unit/spec/Infrastructure/InternalApi/Controller/ErrorManagement/ErrorManagementActionSpec.php
@@ -2,18 +2,18 @@
 
 declare(strict_types=1);
 
-namespace spec\Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller;
+namespace spec\Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\ErrorManagement;
 
 use Akeneo\Connectivity\Connection\Application\ErrorManagement\Query\GetConnectionBusinessErrorsHandler;
 use Akeneo\Connectivity\Connection\Domain\ErrorManagement\Model\Read\BusinessError;
-use Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\ErrorManagementController;
+use Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\ErrorManagement\ErrorManagementAction;
 use PhpSpec\ObjectBehavior;
 
 /**
  * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
  * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class ErrorManagementControllerSpec extends ObjectBehavior
+class ErrorManagementActionSpec extends ObjectBehavior
 {
     public function let(GetConnectionBusinessErrorsHandler $getConnectionBusinessErrorsHandler): void
     {
@@ -22,7 +22,7 @@ class ErrorManagementControllerSpec extends ObjectBehavior
 
     public function it_is_initializable(): void
     {
-        $this->shouldBeAnInstanceOf(ErrorManagementController::class);
+        $this->shouldBeAnInstanceOf(ErrorManagementAction::class);
     }
 
     public function it_normalizes_business_errors(): void
