+++ src/Akeneo/Connectivity/Connection/back/tests/phpstan.neon
@@ -1,4 +1,7 @@
 rules:
+    - Akeneo\Connectivity\Connection\Tests\PhpStan\Rule\FinalControllerRule
+    - Akeneo\Connectivity\Connection\Tests\PhpStan\Rule\InvokableControllerRule
+    - Akeneo\Connectivity\Connection\Tests\PhpStan\Rule\SuffixControllerByActionRule
     - Akeneo\Connectivity\Connection\Tests\PhpStan\Rule\SuffixExceptionRule
     - Akeneo\Connectivity\Connection\Tests\PhpStan\Rule\SuffixInterfaceRule
     - Akeneo\Connectivity\Connection\Tests\PhpStan\Rule\SuffixTraitRule
